pyproject.toml
[project]
name = "mindevolve"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anyio>=4.11.0",
    "dacite>=1.9.2",
    "dotenv>=0.9.9",
    "flask>=3.1.2",
    "importlib>=1.0.4",
    "numpy>=2.3.3",
    "openai>=2.1.0",
    "pandas>=2.3.3",
    "paramiko>=4.0.0",
    "pathlib>=1.0.1",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "scipy>=1.16.2",
    "shutils>=0.1.0",
    "typing>=3.10.0.0",
    "uuid>=1.30",
]
